+++ mangadlp/utils.py
@@ -83,7 +83,6 @@ def get_chapter_list(chapters: str, available_chapters: list) -> list:
 
 # remove illegal characters etc
 def fix_name(filename: str) -> str:
-    log.debug(f"Input name='{filename}'")
     filename = filename.encode(encoding="utf8", errors="ignore").decode(encoding="utf8")
     # remove illegal characters
     filename = re.sub(r'[/\\<>:;|?*!@"]', "", filename)
@@ -94,7 +93,7 @@ def fix_name(filename: str) -> str:
     # remove trailing and beginning spaces
     filename = re.sub("([ \t]+$)|(^[ \t]+)", "", filename)
 
-    log.debug(f"Output name='{filename}'")
+    log.debug(f"Input name='{filename}', Output name='{filename}'")
     return filename